inference/requirements.txt
ultralytics
opencv-python-headless
paho-mqtt
